In-class show. Still need to fix charge.

+++ Assets/MouseLook.cs
@@ -6,7 +6,7 @@ public class MouseLook : MonoBehaviour {
 
 	public GameObject playerBody;
 	static public Vector3[] lineVectors;
-	static public int bounceMax = 2;
+	static public float bounceMax = 2;
 
 	public float horizontalSpeed = 2.0f;
 	public float verticalSpeed = 2.0f;
